+++ dart/lib/sentry.dart
@@ -3,14 +3,17 @@
 // found in the LICENSE file.
 
 /// A pure Dart client for Sentry.io crash reporting.
+export 'src/default_integrations.dart';
+export 'src/hub.dart';
+// useful for tests
+export 'src/hub_adapter.dart';
+export 'src/noop_isolate_error_integration.dart'
+    if (dart.library.io) 'src/isolate_error_integration.dart';
 export 'src/protocol.dart';
 export 'src/scope.dart';
 export 'src/sentry.dart';
 export 'src/sentry_client.dart';
-export 'src/hub.dart';
-// useful for test
-export 'src/hub_adapter.dart';
 export 'src/sentry_options.dart';
-export 'src/transport/transport.dart';
 // useful for integrations
 export 'src/throwable_mechanism.dart';
+export 'src/transport/transport.dart';